+++ src/replica/replica_stub.h
@@ -307,7 +307,7 @@ private:
     friend class replica_duplicator_manager_test;
     friend class duplication_test_base;
     friend class replica_test;
-    friend class replica_disk_test;
+    friend class replica_disk_test_base;
 
     typedef std::unordered_map<gpid, ::dsn::task_ptr> opening_replicas;
     typedef std::unordered_map<gpid, std::tuple<task_ptr, replica_ptr, app_info, replica_info>>
